Fix top navigation bar
Fixed the top nav bar so that user does not scroll to top to navigate
to a different page. Changed id of header from caption to captionFeatures,
contentHolder to contentHolderFeatures to fix features.jsp.
This commit modifies paddings and margins in teammates.css to fix
the navigation bar on the top.